Renamed mobile-aisleriot to patience-collection
This is done to avoid confusion about the relationship of this project
and GNOME Aisleriot. Thus rename all instances of mobile-aisleriot to
patience-collection and Aisleriot class to Patience.

Also decided to rename Board class to Table as it seems more suitable
name for the purpose and at the same go changed how paths are specified
in code.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>