Add allowlist for supported games
Add allowlist for supported games to only show those by default. It can
be changed on About page. For now it's volatile and forgotten when the
application is restarted but it will be backed up by a dconf value
in future.

Currently only listing four games there but it can be expanded once more
of them are tested to be working. Also even these games have some
problems at the moment but the aim is to fix them for the first release.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>