New toolbar design
This has now expand button to show more buttons. On portrait it's
positioned the same as before but on landscape it's on the left side.
Expanding it will move the game to the right. This will affect
performance a little on slower devices but not much. New deal and hint
buttons are not yet functional.

Add graphics for expand button.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>