Drop to the area of the dragged card
Previously overlap was calculated using the whole stack of cards that
was dragged but now it's calculated only for the card user started
dragging from. It's likely what user intended to do and allows to finish
Spider again.

Also use mapToItem() to calculate position relative to Table so it's
meaningful to compare it between different mouse events.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>