New icon
Better design and automatic conversion.

The conversion code is free to use for any purpose. It's also so short
that it would be hard to call it really copyrightable although I did
write it from scratch. I added small 0BSD snippet for license so there
shouldn't be any question about that.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>