Shrink toolbar when transitioning to a subpage
This prevents drawing table over the subpage. It's really a workaround
because I could not figure out in time how to make the clipping behave
nicely. Currently the whole Table item just disappears when clipping is
active. This would also allow me to keep Table centered in
tableContainer and clip the sides when toolbar is expanded.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>