Fix up game status and timer handling
Prevent pausing or unpausing timer when game is not running.

Rename RestartedState to StartedState and use it at the beginning of
every game.

Move to RunnigState only when the very first move is made which matches
the behaviour of GNOME Aisleriot and many other patience games.

Also ensure that once timer is stopped it can't be unpaused by using
stop and extend instead of pause and unpause.

When returning to application, unpause only if the page is active.
Also don't treat PageStatus.Activating as active state.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>