Add card highlighting support
Allow highlighting cards and slots that dragged cards can be dropped to.
While dragging or dropping go through the possible slots and check if
they can be used for dropping. If there is a cache miss then check the
engine and after that returns, continue the process.

Make Drag QQuickItem. This way it can be parentItem for Cards while they
are dragged which simplifies the code.

Add DragState enum that allows better track how drag is progressing and
replaces some boolean values.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>