Add engine-exerciser tool
Currently this just allows to run the engine without the app to play
through games of Klondike automatically. It uses hints to figure out the
next moves which allows to solve some of the games. While it often
doesn't finish games it still allows sufficient amount of testing for
the engine's functions.

Engine was adjusted so that it builds without mlite and other parts that
may store or load some state. Without that the engine compiles fine
also outside Sailfish SDK. EngineHelper class implements testing related
functionality that isn't part of Engine class itself.

This is very "bare bones" version that doesn't implement anything extra
like choosing which game is loaded or selecting seed.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>