Reverse card removal order in engine
Iterate slots backwards when adjusting them. This fixes a corner case in
Hamilton and hopefully doesn't regress other games. Cards are usually
taken from top of the slots so this should work nicely.

The bug in Hamilton could be reproduced like this:
- Start a new game of Hamilton
- Touch stock twice
- Double tap the card in waste to select it
- Deal a new round

It should deal new cards but instead manager is stuck and can't put the
cards to their slots. This happened because the engine emptied the stock
just to add one card there. As one of the cards was removed and inserted
to two different places on the same move, manager happened to pick the
wrong insertion first which deadlocked one of the slots and manager
could not advance anymore. As far as I know there is no other way to
reproduce this, just dragging that card would make it work fine and
interestingly selecting the third card doesn't have this problem.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>