Render card texture in another thread
Move card size computation to updatePolish() to avoid calculating it
many times. This avoids rendering the texture more than once when
multiple affecting properties have been changed. This is important since
rendering textures in a second thread is not useful if they need to be
rendered multiple times in a row.

After this rotating device will not draw table first in wrong
orientation and then in the correct orientation. Instead it will first
use the old texture to draw the cards and then a moment later draw the
cards with the correct texture.

Set composition mode correctly and fill the texture with transparent
color effectively clearing the buffer before rendering SVG. That fixes a
problem with card corners getting garbled on some devices after a few
display rotations.

Since the resulting QImage is now effectively cached, leaving the app
and returning back to it is now a little bit faster too.

Also mark child QSGNodes owned by their parents.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>