Fix returning to game with lots of radio options
If there are lots of radio options, they get pushed to a new page. In
that case it didn't pop the correct pages from page stack.

Unfortunately this does complain about popping while transition is
ongoing but it seems that I can't avoid it. ComboBox doesn't respect
ContextMenu.closeOnActivation, provide anything similar of its own or
does it provide any mechanism to select which page it should pop to.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>