Make action queue its own class
Separate Queue class from Manager. This should not change behaviour,
just separate the code in preparation for the changes in future.

This also brings back delaying all actions until moveEnded signal is
received. Otherwise the intermediate state may be shown to user even if
that's not intended.

Also changed CardData to return SuitAndRank with value().

Signed-off-by: Tomi Leppänen <tomi@tomin.site>