Don't miss StartingDrag state in cancel()
When drag has been started and it's immediately ended before couldDrag
signal is received, then the drag must be canceled in case it was able
to begin. If it wasn't then the engine will print a warning that can be
ignored. This avoids leaving engine in state where it thinks there is an
action ongoing but UI side disagrees.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>