Allow canceling Drag only before calling drop
Add check to act on cancel() only before drop() is sent to engine. This
avoids another touch canceling Drag before engine has sent dropped()
back to Drag. Rename Dropped state to Finished and add Dropped between
Dropped and Finished to have a state for the period between having sent
drop() but not yet received dropped(). This is similar to states with
other signals used to communicate with the engine.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>