When unable to drag, handle it gracefully
Fix situations where drag is not possible. Previously this resulted in
likely harmless errors. Also fix a bit the engine side as we should
always also end or discard moves that are started.

This was a simple omission on couldDrag handling that false case didn't
result in canceling move. Additionally make finish() to destroy the drag
instance because we want it to eat all of the mouse events.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>