Don't draw cards below the topmost non-expanded card
This is an optimization. When enabling overdraw visualization, it's
clear that there are cards on top of each other that could be not drawn
at all. As it's fairly simple to do in updateLocations() and apart from
needing to figure out that firstExpanded iterator mostly cheap, let's do
that.

Also fix firstExpanded() for non-expanded slots.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>