Fix overlap calculation
For some reason now if I moved a card from a slot to another in Spider
Three Decks, it can leave the old slot's childrenRect too big which
breaks overlap calculation in Table::getSlotsFor(). This improved slot
box calculation avoids using childrenRect and it builds on the
assumption that cards are placed to slot in order which should be always
true.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>