Allow table clipping on landscape orientation
This was an old bug originating from the toolbar redesign. Apparently I
misspelled isLandscape and that caused issues I didn't understand. Fix
them and make table centered on horizontal direction. This allows
playing while toolbar is expanded. Also allows to remove that workaround
that toolbar is closed when subpages are opened. That's just not
necessary anymore.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>