Improved toolbar
This version has a handle which can be dragged to open the toolbar. It
is now possible to fit all buttons on the first row when there is space
on landscape. On portrait all buttons are on the same row but they are
now draggable to expose more buttons.

There are some layers of clipping which probably could be reduced and
there is also a lot of animations some of which could be maybe replaced
with animators.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>