Fix haptic feedback with selections
Previously double tap on card would cause a bit strange vibration to
occure as it first vibrates on failed drop and then again on successful
drop. With this the first vibration is skipped.

Add selectionChanged signal to Feedback. This is played when selection
changes but no cards are moved. In practice it replaces dropFailed for
selections because any failure will remove selection without moving
cards. The feedback effect defined for this is the same short vibration.
Drop sets this vibration blocked because it overrides the vibration
handling.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>