Fix icon path resolution if there is no exact match
Previously no icon was found if device did not use one of the three icon
sizes. Fix this by figuring out the closest match, but preferring icons
larger than the perfect match. This should not change behaviour for
devices that use the sizes that already exist.

Also add that missing logging category for critical error.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>