Add MoveEndedAction
This is important for future development. It makes sure that all actions
have been processed before end of move is handled in queue. It also
makes it possible to have different conditions for move ended signaling
when actions need to be played through and when there is actually
nothing happening in the engine (moveEnded() signal).

Signed-off-by: Tomi Leppänen <tomi@tomin.site>