Add testing script and test data
Add testing script and test data for currently supported games. The test
cases are of varying quality but given that most of the features are
shared between different games it probably gives fairly good coverage.

Note that the tests consider any kind of warning as a bug and thus some
of the tests fail even if there is nothing wrong with replaying the
games. Those should be still considered as bugs and should be fixed.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>