Add patch to adjust Clock for animations
Clock has extended slots that never have more than four cards but their
extension is not limited which means that dropped card tries to move to
fifth place in that slot and then its immediately moved to the fourth
position when the first card is removed from that slot. This looks
erratic so let's fix it with a patch. Aisleriot does not support
animations so upstream is not concerned about this.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>