Fix button behaviour
Force consistent button order. Previously deal and restart buttons
switched places if you rotated from portrait to landscape and back.
Force parent evaluation of buttons to happen in order.

Fix showing restart button when there are four buttons and they all can
fit to toolbar. For example Spider on tablet device. Before 5c93a43 this
did not matter as much as it would now that the handle is not shown in
this case.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>