Add support for colouring textures
Add SvgDocument class that inherits QBuffer. It first loads the content
to DOM and sets CSS styles based on class names. Then it fills the
buffer with that content so that QSvgRenderer can consume it.

Read colours from config in TextureRenderer. Add PerfTimer class to
measure texture loading and rendering performance.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>