Add support for 'I', 'L' and 'T' commands
Add 'I' command that can 'isolate' elements from the element they link
to. This is important if you want to modify either of these elements
without affecting the other. It also takes class names as an argument.

Add 'L' command that can set class names. 'L' is the only letter in word
class that has not yet been used as a command.

Add 'T' command to card style converter to set style. ;) It's also T for
'tyyli' ('style' in Finnish).

Signed-off-by: Tomi Leppänen <tomi@tomin.site>