Use double resolution texture while magnifying
Use double the resolution texture while magnifying if the cards are
small. This fixes blurry cards when magnifying on games that have really
small cards. As very large textures have poor performance, only games
that have small enough cards use double resolution. Bigger cards look
anyway good when magnified.

The limit for double resolution is a combination of a size that's known
to work and some experimentation to drive it slightly higher without
sacrificing performance. With the largest cards texture upload can take
a good while if double resolution was used, so this is only slightly
larger than the largest texture before.

This does slow down magnifying a little. There is a slight delay before
it magnifies when the texture is large.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>