Allow keeping the state for the previous game
This implementes enablers for new functionality for restoring the
previously played game after starting a new game. The old state can be
forgotten or restored from QML.

Setting game options or changing to another game invalidates old state
because it's just not compatible with this system. In theory those could
also be made to work but remorse timer that delays the action might
serve the user better there.

Restoring previous state is prevented while engine is busy. If a move is
ongoing or a game is being replayed, the engine should not attempt to
restore previous game. This tries to avoid situations where player is
quickly moving between previous and new game that could otherwise result
in invalid state or even outright crash.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>