New graphics for restart, new and restore buttons
Update restart button to have a few cards at the center. Add new game
and restore button graphics that have the same symbol. This should
provide a strong connection between the actions that these buttons
provide: restart restores current shuffle from the beginning of the game
and new game reshuffles cards. Restore returns to the previous shuffling
of the cards.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>