Add animation builder
It can do animations that move object to any direction. Either in xy
coordinates or a length to specified direction. There is also a helper
to create xy animation for one object which is used for positioning Drag
after dropping. It can also change z order.

While the builder can make animations that run in sequence or in
parallel. To add parallel animations to sequential animation, use
ParallelGroupGuard. There isn't similar guard to group sequential
animations but it could be added.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>