Store RNG state with recorder
Change Recorder to store RNG state so it can correctly save also games
like Gaps that allow shuffling cards during game play. This would be
only a problem when undo clears shuffling and randomness is used again
later. After this change we remember if any undo'd action advanced RNG
state and we can take that into account during replay. Previously
restoring such games always failed or had an unexpected result.

Note that this adds data version "1" which is not compatible with data
version "0". Older versions of the game will refuse to replay games that
are stored with this new data version.

This change is necessary for games that shuffle cards during game play.
The new version will be used only after that happens. Technically we
would need it only if player also used undo over a move that shuffled
the cards but this code inserts RNG state and upgrades data version
every time cards are shuffled after game setup for a simpler
implementation.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>