Add CRC checksum to saved state
Store CRC checksum of topmost cards as part of saved state. The
intention is that in case the restoration does not fail, this would
detect any changes in the end result. This is mostly useful for
automated tests but it could save some corner case for user too when the
game files are updated some day. The saved checksum is versioned so that
it can be changed independently of the rest of the state.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>