Update app cover
Improve skull icon. Remove the bones and add teeth. Make eye holes
round. Add a little bit of transparency.

Add transparency to fireworks icon. This adds a bit of depth and makes
it look more interesting. Also the cover demands a little bit of
transparency.

Add cards cover icon. This displays when game is still running or has
not started yet. The icon has Q, K and A cards on top of each other a
little bit offset and rotated. The style of the cards is loosely based
on the anglo card set: spade shape and letter shapes are the same and
the outline of the cards and partially the king shape is used.

Update app cover to always show one of these icons instead of the app
icon. This makes the cover nicer looking and more interesting.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>